requirements.txt
aiogram==3.4.1
wikipedia==1.4.0
python-dotenv==1.0.1
